python3/requirements.txt
defusedxml==0.4.1
Django==1.8.4
djangorestframework==3.2.2
djangorestframework-xml==1.2.0
wheel==0.24.0
